+++ Modules/lootFrame.lua
@@ -92,8 +92,9 @@ function LootFrame:Update()
 			entries[numEntries].icon:SetNormalTexture(v.texture)
 			entries[numEntries].itemText:SetText(v.link)
 			entries[numEntries].itemLvl:SetText(format(L["ilvl: x"], v.ilvl))
-			if not entries[numEntries].buttons[addon.mldb.numButtons+1] then
+			if not entries[numEntries].buttons[addon.mldb.numButtons+1] or (entries[numEntries].buttons[addon.mldb.numButtons+1]:GetText() ~= L["Pass"]) then
 				 -- mldb have probably updated since we created the buttons
+				 addon:Debug("Updating buttons")
 				 entries[numEntries]:UpdateButtons()
 			end
 			-- Update the buttons and get frame width
@@ -193,9 +194,13 @@ function LootFrame:GetEntry(entry)
 			f.buttons[i]:SetScript("OnClick", function() LootFrame:OnRoll(entry, i) end)
 		end
 		-- Pass button
-		f.buttons[addon.mldb.numButtons + 1] = addon:CreateButton(L["Pass"], f)
+		f.buttons[addon.mldb.numButtons + 1] = f.buttons[addon.mldb.numButtons + 1] or addon:CreateButton(L["Pass"], f)
+		f.buttons[addon.mldb.numButtons + 1]:SetText(L["Pass"])
 		f.buttons[addon.mldb.numButtons + 1]:SetPoint("LEFT", f.buttons[addon.mldb.numButtons], "RIGHT", 5, 0)
 		f.buttons[addon.mldb.numButtons + 1]:SetScript("OnClick", function() LootFrame:OnRoll(entry, "PASS") end)
+		if #f.buttons > addon.mldb.numButtons + 1 then
+			for i = addon.mldb.numButtons +2, #f.buttons do f.buttons[i]:Hide() end
+		end
 	end
 	f:UpdateButtons()